# contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Accounting API",
    "version": "0.2.0",
    "description": "Stage 1 manual accounting contract. Future imports, budgets, reports and sync transport are specified separately; no server exists yet."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/workspaces/{workspace_id}/sync/snapshots": {
      "post": {
        "operationId": "sync_snapshot_create",
        "summary": "sync snapshot create",
        "tags": [
          "sync"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/WorkspaceId"
          },
          {
            "$ref": "#/components/parameters/ActionId"
          }
        ],
        "description": "Read-only materialization, Idempotency-Key must equal body id. Retries reuse same actor/workspace/id snapshot while live. No journal group, not a financial actions receipt. Expired id => 410 snapshot_expired; use new snapshot id.",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "410": {
            "$ref": "#/components/responses/Error410"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotCreate"
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspace_id}/sync/snapshots/{id}": {
      "get": {
        "operationId": "sync_snapshot_get",
        "summary": "sync snapshot get",
        "tags": [
          "sync"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/WorkspaceId"
          },
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "description": "Actor-scoped immutable snapshot metadata; TTL 15 minutes. Reauthorize membership on every read.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "410": {
            "$ref": "#/components/responses/Error410"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        }
      }
    },
    "/workspaces/{workspace_id}/sync/snapshots/{id}/pages": {
      "get": {
        "operationId": "sync_snapshot_page",
        "summary": "sync snapshot page",
        "tags": [
          "sync"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/WorkspaceId"
          },
          {
            "$ref": "#/components/parameters/ResourceId"
          },
          {
            "name": "page_token",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageToken"
            }
          }
        ],
        "description": "Fixed pages of at most 100 items. Repeat same token returns same page while snapshot lives. Install staged mirror and resume cursor only after final page.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "410": {
            "$ref": "#/components/responses/Error410"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UUID": {
        "type": "string",
        "format": "uuid"
      },
      "Version": {
        "type": "string",
        "pattern": "^[1-9][0-9]{0,18}$",
        "description": "Positive int64 encoded as a string; server checks <= 9223372036854775807."
      },
      "Money": {
        "type": "string",
        "pattern": "^(0|-?([1-9][0-9]{0,14}|[1-8][0-9]{15}|9000000000000000))$",
        "description": "Signed minor units, absolute value <= 9000000000000000; encoded as a canonical decimal integer string."
      },
      "PositiveMoney": {
        "type": "string",
        "pattern": "^([1-9][0-9]{0,14}|[1-8][0-9]{15}|9000000000000000)$",
        "description": "Positive minor units <= 9000000000000000."
      },
      "PositiveInteger": {
        "type": "string",
        "pattern": "^[1-9][0-9]*$",
        "maxLength": 40,
        "description": "Exact positive numerator/denominator; no float."
      },
      "Instant": {
        "type": "string",
        "format": "date-time",
        "pattern": "Z$",
        "description": "UTC RFC3339 instant ending in Z."
      },
      "Timezone": {
        "type": "string",
        "minLength": 1,
        "maxLength": 100,
        "description": "IANA timezone; checked against timezone database."
      },
      "CurrencyCode": {
        "type": "string",
        "enum": [
          "EUR",
          "USD",
          "RUB",
          "GBP",
          "JPY",
          "KWD"
        ]
      },
      "Rate": {
        "type": "object",
        "properties": {
          "numerator": {
            "$ref": "#/components/schemas/PositiveInteger"
          },
          "denominator": {
            "$ref": "#/components/schemas/PositiveInteger"
          }
        },
        "required": [
          "numerator",
          "denominator"
        ],
        "additionalProperties": false
      },
...
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "workspace_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "created_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "updated_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "deleted_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instant"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "enum": [
              "cash",
              "bank",
              "card"
            ]
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyCode"
          },
          "opened_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "archived_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instant"
              },
              {
                "type": "null"
              }
            ]
          },
          "balance_version": {
            "$ref": "#/components/schemas/Version"
          },
          "posted_balance_minor": {
            "$ref": "#/components/schemas/Money"
          },
          "pending_delta_minor": {
            "$ref": "#/components/schemas/Money"
          },
          "projected_balance_minor": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "version",
          "created_at",
          "updated_at",
          "deleted_at",
          "name",
          "type",
          "currency",
          "opened_at",
          "archived_at",
          "balance_version",
          "posted_balance_minor",
          "pending_delta_minor",
          "projected_balance_minor"
        ],
        "additionalProperties": false
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "workspace_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "created_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "updated_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "deleted_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instant"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "archived_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instant"
              },
              {
                "type": "null"
              }
            ]
          },
          "parent_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UUID"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "workspace_id",
          "version",
          "created_at",
          "updated_at",
          "deleted_at",
          "name",
          "archived_at",
          "parent_id"
        ],
        "additionalProperties": false
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "workspace_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "created_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "updated_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "deleted_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instant"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "archived_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instant"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "workspace_id",
          "version",
          "created_at",
          "updated_at",
          "deleted_at",
          "name",
          "archived_at"
        ],
        "additionalProperties": false
      },
...
      "Entry": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "account_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyCode"
          },
          "amount_minor": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "required": [
          "id",
          "account_id",
          "currency",
          "amount_minor"
        ],
        "additionalProperties": false
      },
...
      "Allocation": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "category_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UUID"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount_minor": {
            "$ref": "#/components/schemas/PositiveMoney"
          },
          "original_allocation_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UUID"
              },
              {
                "type": "null"
              }
            ]
          },
          "remaining_refundable_minor": {
            "$ref": "#/components/schemas/Money",
            "description": "Nonnegative remaining amount only for allocations of expense; zero for all other kinds. Active posted/pending refunds reserve the limit."
          }
        },
        "required": [
          "id",
          "category_id",
          "amount_minor",
          "original_allocation_id",
          "remaining_refundable_minor"
        ],
        "additionalProperties": false
      },
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "workspace_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "created_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "updated_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "deleted_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instant"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string",
            "enum": [
              "opening",
              "expense",
              "income",
              "transfer",
              "refund",
              "adjustment"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "posted"
            ]
          },
          "occurred_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "occurred_timezone": {
            "$ref": "#/components/schemas/Timezone"
          },
          "note": {
            "type": "string",
            "maxLength": 2000
          },
          "payee": {
            "type": "string",
            "maxLength": 200
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ]
          },
          "parent_transaction_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UUID"
              },
              {
                "type": "null"
              }
            ]
          },
          "fee_transaction_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UUID"
              },
              {
                "type": "null"
              }
            ]
          },
          "rate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Rate"
              },
              {
                "type": "null"
              }
            ]
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Entry"
            },
            "maxItems": 2
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Allocation"
            },
            "maxItems": 100
          },
          "tag_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UUID"
            },
            "uniqueItems": true,
            "maxItems": 50
          }
        },
        "required": [
          "id",
          "workspace_id",
          "version",
          "created_at",
          "updated_at",
          "deleted_at",
          "kind",
          "status",
          "occurred_at",
          "occurred_timezone",
          "note",
          "payee",
          "reason",
          "parent_transaction_id",
          "fee_transaction_id",
          "rate",
          "entries",
          "allocations",
          "tag_ids"
        ],
        "additionalProperties": false
      },
...
      "FieldError": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "code",
          "message"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "request_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "code": {
            "type": "string",
            "enum": [
              "bad_request",
              "validation_error",
              "unauthenticated",
              "forbidden",
              "csrf_invalid",
              "not_found",
              "version_conflict",
              "idempotency_conflict",
              "dependent_transactions",
              "account_archived",
              "amount_out_of_range",
              "no_change",
              "rate_limited",
              "service_unavailable",
              "internal_error",
              "cursor_expired",
              "cursor_invalid",
              "snapshot_expired",
              "snapshot_too_large",
              "action_result_expired",
              "sync_group_too_large",
              "sync_generation_conflict",
              "registration_disabled"
            ]
          },
          "message": {
            "type": "string"
          },
          "field_errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "current_versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConflictVersion"
            }
          },
          "retry_after_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          }
        },
        "required": [
          "request_id",
          "code",
          "message",
          "field_errors",
          "current_versions",
          "retry_after_seconds"
        ],
        "additionalProperties": false
      },
...
      "ConflictVersion": {
        "type": "object",
        "properties": {
          "entity_type": {
            "type": "string",
            "enum": [
              "account",
              "transaction",
              "category",
              "tag"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "balance_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Version"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "entity_type",
          "id",
          "version",
          "balance_version"
        ],
        "additionalProperties": false
      },
      "Sequence": {
        "type": "string",
        "pattern": "^(0|[1-9][0-9]{0,18})$",
        "description": "Committed workspace group sequence, zero for empty workspace; service checks int64 range."
      },
      "SyncCursor": {
        "type": "string",
        "minLength": 16,
        "maxLength": 2048,
        "description": "Opaque integrity-protected workspace/actor/protocol-bound journal cursor. Not a list pagination cursor."
      },
      "PageToken": {
        "type": "string",
        "minLength": 16,
        "maxLength": 2048,
        "description": "Opaque snapshot-specific token. Immutable page position; not a sync cursor."
      },
...
      "SnapshotItem": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "entity_type": {
                "const": "account"
              },
              "data": {
                "$ref": "#/components/schemas/Account"
              }
            },
            "required": [
              "entity_type",
              "data"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "entity_type": {
                "const": "transaction"
              },
              "data": {
                "$ref": "#/components/schemas/Transaction"
              }
            },
            "required": [
              "entity_type",
              "data"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "entity_type": {
                "const": "category"
              },
              "data": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "required": [
              "entity_type",
              "data"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "entity_type": {
                "const": "tag"
              },
              "data": {
                "$ref": "#/components/schemas/Tag"
              }
            },
            "required": [
              "entity_type",
              "data"
            ],
            "additionalProperties": false
          }
        ]
      },
      "SnapshotCreate": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          }
        },
        "required": [
          "id"
        ],
        "additionalProperties": false
      },
      "Snapshot": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "workspace_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "base_sequence": {
            "$ref": "#/components/schemas/Sequence"
          },
          "resume_cursor": {
            "$ref": "#/components/schemas/SyncCursor"
          },
          "created_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "expires_at": {
            "$ref": "#/components/schemas/Instant"
          },
          "item_count": {
            "type": "integer",
            "minimum": 0,
            "maximum": 50000
          },
          "first_page_token": {
            "$ref": "#/components/schemas/PageToken"
          },
          "generation_id": {
            "$ref": "#/components/schemas/UUID"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "base_sequence",
          "resume_cursor",
          "created_at",
          "expires_at",
          "item_count",
          "first_page_token",
          "generation_id"
        ],
        "additionalProperties": false
      },
      "SnapshotPage": {
        "type": "object",
        "properties": {
          "snapshot_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "workspace_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "base_sequence": {
            "$ref": "#/components/schemas/Sequence"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SnapshotItem"
            },
            "maxItems": 100
          },
          "next_page_token": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageToken"
              },
              {
                "type": "null"
              }
            ]
          },
          "generation_id": {
            "$ref": "#/components/schemas/UUID"
          }
        },
        "required": [
          "snapshot_id",
          "workspace_id",
          "base_sequence",
          "items",
          "next_page_token",
          "generation_id"
        ],
        "additionalProperties": false
      },
...
    },
    "parameters": {
      "WorkspaceId": {
        "name": "workspace_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/UUID"
        }
      },
      "ResourceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/UUID"
        }
      },
      "ActionId": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/UUID"
        },
        "description": "Actor/workspace scoped action UUID; retry unchanged request with same key."
      },
...
    },
    "responses": {
      "Error400": {
        "description": "Malformed request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error401": {
        "description": "Missing, expired or revoked session",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error403": {
        "description": "CSRF check failed or operation forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error404": {
        "description": "Resource missing or inaccessible",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error409": {
        "description": "Version, idempotency or dependency conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "Idempotency-Replayed": {
            "description": "Present when returning a registered terminal action result; true on retained replay, false on initial completion. Not set for nonregistered errors.",
            "schema": {
              "type": "boolean"
            }
          }
        }
      },
      "Error422": {
        "description": "Domain validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "Idempotency-Replayed": {
            "description": "Present when returning a registered terminal action result; true on retained replay, false on initial completion. Not set for nonregistered errors.",
            "schema": {
              "type": "boolean"
            }
          }
        }
      },
      "Error429": {
        "description": "Rate limited",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Seconds before retry"
          }
        }
      },
      "Error500": {
        "description": "Internal error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error503": {
        "description": "Temporary unavailability",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Seconds before retry"
          }
        }
      },
      "Error413": {
        "description": "Request body exceeds configured limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error410": {
        "description": "Sync cursor or snapshot expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque session token, not JWT."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-accounting_session"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    }
  }
}